feat(component): refactor CellSizeSelector component to use Dropdown and simplify button rendering
