+++ src/components/CellSizeSelector/CellSizeSelector.tsx
@@ -1,32 +1,21 @@
-import { TCellSizeSelectorProps } from "./CellSizeSelector.type";
+import { Dropdown } from "@/components/UI";
 import { useTranslation } from "react-i18next";
-import { getButtonClasses } from "./CellSizeSelector.util";
+import type { TCellSizeSelectorProps } from "./CellSizeSelector.type";
 
 export function CellSizeSelector({ activeSize, options, onSelect }: TCellSizeSelectorProps) {
   const { t } = useTranslation();
 
   return (
-    <div className={`flex items-center gap-2 md:gap-4 flex-wrap`}>
-      <span
-        className={`
-          text-[length:var(--font-md)] text-[var(--color-text-strong)] 
-          whitespace-nowrap
-        `}
-      >
+    <div className="flex flex-col gap-1.5">
+      <p className="mb-2 text-xs font-semibold uppercase tracking-widest text-[var(--color-text-secondary)]">
         {t("climateStatistics.cellResolution")}
-      </span>
-
-      <div className={`flex gap-3`}>
-        {options.map((option) => (
-          <button
-            key={option.value}
-            onClick={() => onSelect(option.value)}
-            className={getButtonClasses(activeSize, option.value)}
-          >
-            {option.label}
-          </button>
-        ))}
-      </div>
+      </p>
+      <Dropdown
+        options={options.map((o) => ({ value: o.value, label: o.label }))}
+        value={activeSize}
+        onChange={(v) => onSelect(v as typeof activeSize)}
+        className="w-full"
+      />
     </div>
   );
 }
